moved media stuff to delegate
