+++ feature/post/src/main/java/org/mozilla/social/post/NewPostViewModel.kt
@@ -1,9 +1,7 @@
 package org.mozilla.social.post
 
-import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
@@ -19,17 +17,15 @@ import org.mozilla.social.common.LoadState
 import org.mozilla.social.common.logging.Log
 import org.mozilla.social.core.data.repository.MediaRepository
 import org.mozilla.social.core.data.repository.StatusRepository
-import org.mozilla.social.model.ImageState
-import org.mozilla.social.model.entity.PollOption
 import org.mozilla.social.model.entity.StatusVisibility
 import org.mozilla.social.model.entity.request.PollCreate
 import org.mozilla.social.post.contentwarning.ContentWarningDelegate
-import org.mozilla.social.post.interactions.ContentWarningInteractions
-import org.mozilla.social.post.interactions.ImageInteractions
-import org.mozilla.social.post.interactions.PollInteractions
+import org.mozilla.social.post.contentwarning.ContentWarningInteractions
+import org.mozilla.social.post.media.MediaInteractions
+import org.mozilla.social.post.poll.PollInteractions
+import org.mozilla.social.post.media.MediaDelegate
 import org.mozilla.social.post.poll.PollDelegate
 import org.mozilla.social.post.poll.PollStyle
-import java.io.File
 
 class NewPostViewModel(
     private val statusRepository: StatusRepository,
@@ -38,19 +34,25 @@ class NewPostViewModel(
     private val onStatusPosted: () -> Unit,
     private val pollDelegate: PollDelegate = PollDelegate(),
     private val contentWarningDelegate: ContentWarningDelegate = ContentWarningDelegate(),
+    private val mediaDelegate: MediaDelegate = MediaDelegate(
+        mediaRepository,
+        log,
+    ),
 ) : ViewModel(),
-    ImageInteractions,
+    MediaInteractions by mediaDelegate,
     PollInteractions by pollDelegate,
     ContentWarningInteractions by contentWarningDelegate
 {
+    init {
+        mediaDelegate.coroutineScope = viewModelScope
+    }
+
     val poll = pollDelegate.poll
     val contentWarningText = contentWarningDelegate.contentWarningText
+    val imageStates = mediaDelegate.imageStates
 
     private val _statusText = MutableStateFlow("")
     val statusText = _statusText.asStateFlow()
-
-    private val _imageStates = MutableStateFlow<Map<Uri, ImageState>>(emptyMap())
-    val imageStates = _imageStates.asStateFlow()
 
     val sendButtonEnabled: StateFlow<Boolean> =
         combine(statusText, imageStates, poll) { statusText, imageStates, poll ->
@@ -92,8 +94,6 @@ class NewPostViewModel(
     private val _visibility = MutableStateFlow(StatusVisibility.Public)
     val visibility = _visibility.asStateFlow()
 
-    private val uploadJobs = mutableMapOf<Uri, Job>()
-
     fun onStatusTextUpdated(text: String) {
         if (text.length + (contentWarningText.value?.length ?: 0) > MAX_POST_LENGTH) return
         _statusText.update { text }
@@ -102,58 +102,6 @@ class NewPostViewModel(
     override fun onContentWarningTextChanged(text: String) {
         if (text.length + statusText.value.length > MAX_POST_LENGTH) return
         contentWarningDelegate.onContentWarningTextChanged(text)
-    }
-
-    override fun onImageDescriptionTextUpdated(
-        uri: Uri,
-        text: String,
-    ) {
-        if (text.length > MAX_POST_LENGTH) return
-        updateImageState(uri, description = text)
-    }
-
-    override fun onDeleteImageClicked(uri: Uri) {
-        uploadJobs[uri]?.cancel()
-        _imageStates.update {
-            _imageStates.value.toMutableMap().apply { remove(uri) }
-        }
-    }
-
-    /**
-     * When an image is inserted, we need to upload it and hold onto the attachment id we get
-     * from the server.
-     */
-    override fun onImageInserted(
-        uri: Uri,
-        file: File,
-    ) {
-        // if the image was already uploaded, just return
-        imageStates.value[uri]?.let {
-            if (it.loadState == LoadState.LOADED) {
-                return
-            }
-        }
-        updateImageState(uri, loadState = LoadState.LOADING)
-        uploadJobs[uri] = viewModelScope.launch {
-            try {
-                val imageId = mediaRepository.uploadImage(
-                    file,
-                    imageStates.value[uri]?.description?.ifBlank { null }
-                ).attachmentId
-                updateImageState(
-                    uri,
-                    loadState = LoadState.LOADED,
-                    attachmentId = imageId
-                )
-            } catch (e: Exception) {
-                log.e(e)
-                updateImageState(uri, loadState = LoadState.ERROR)
-            }
-        }
-        uploadJobs[uri]?.invokeOnCompletion {
-            log.d("removed")
-            uploadJobs.remove(uri)
-        }
     }
 
     fun onVisibilitySelected(statusVisibility: StatusVisibility) {
@@ -187,26 +135,6 @@ class NewPostViewModel(
         }
     }
 
-    private fun updateImageState(
-        uri: Uri,
-        loadState: LoadState? = null,
-        attachmentId: String? = null,
-        description: String? = null,
-    ) {
-        val oldState = _imageStates.value[uri]
-        val newState = ImageState(
-            loadState = loadState ?: oldState?.loadState ?: LoadState.LOADING,
-            attachmentId = attachmentId ?: oldState?.attachmentId,
-            description = description ?: oldState?.description ?: "",
-        )
-
-        _imageStates.update {
-            _imageStates.value.toMutableMap().apply {
-                put(uri, newState)
-            }
-        }
-    }
-
     companion object {
         /**
          * The maximum number of images allowed to be attached to a single post.
@@ -216,5 +144,6 @@ class NewPostViewModel(
         const val MAX_POST_LENGTH = 500
         const val MAX_POLL_COUNT = 4
         const val MIN_POLL_COUNT = 2
+        const val MAX_IMAGE_DESCRIPTION_LENGTH = 1_500
     }
 }
